Fix chained BIO repeat management
Looking at the BIO_CHAIN flag it seems that it appear in the bio after
splitting it. Below is presented real life example. First bio with timestamp
78386.890436 is splitted into two: first with 2048 sectors which is managed
by the Linux kernel and second with timestamp 78387.427706 which size is
equal to 6144 sectors. The second is send to our driver with BIO_CHAIN flag.
Next this bio is splitted into two: first with 2048 sectors is managed by
kernel and second with timestamp 78387.693920 and size 4096 sectors is send
to our driver with BIO_CHAIN flag. So one can see the pattern. First not
fragmented bio is managed by our driver so we do not need to care about
fragments with BIO_CHAIN flag set.

[78386.890436] datto: REQ_OP_WRITE       <>     BIO flags: BIO_TRACE_COMPLETION BIO_REMAPPED    BLK status: BLK_STS_OK   s: 34069984-34078176 8192
[78387.427706] datto: REQ_OP_WRITE       <>     BIO flags: BIO_CHAIN BIO_TRACE_COMPLETION BIO_REMAPPED  BLK status: BLK_STS_OK   s: 34072032-34078176 6144
[78387.693920] datto: REQ_OP_WRITE       <>     BIO flags: BIO_CHAIN BIO_TRACE_COMPLETION BIO_REMAPPED  BLK status: BLK_STS_OK   s: 34074080-34078176 4096
[78388.159220] datto: REQ_OP_WRITE       <>     BIO flags: BIO_CHAIN BIO_TRACE_COMPLETION BIO_REMAPPED  BLK status: BLK_STS_OK   s: 34076128-34078176 2048

The solution that will solve some issues related with bio looping inside the
dattobd include omitting requests with BIO_CHAIN flag set. As described above,
the bio which include chained fragments is managed entirely (all sectors) at
once in dattobd so we do not care about fragments. If we care then it will
make a disaster on some distributions because of too big amount of memory
usage.

The DATTOBD_PASSTHROUGH flag has been used to mark already managed bio. When
one distinguish if the bio need to be managed by the COW manager using
BIO_CHAIN flag then there is no purpose to use this flag. Additionally
the system enum type has been used to mark bio with DATTOBD_PASSTHROUGH
which seems to be not a good solution.